requirements.txt
anthropic>=0.40.0
pybambu>=0.2.0
paho-mqtt>=2.0.0
python-dotenv>=1.0.0
